chore(deps): update kibana-openapi-spec digest to 7bc0014 (#4385)
* chore(deps): update kibana-openapi-spec digest to 7bc0014

* fix(kbapi): adapt transforms for the updated Kibana OpenAPI spec

The new spec already emits dashboard panel oneOf plus discriminator and
changed Lens/control unions. Update schema transforms and provider
converters so renovate-post-upgrade can regenerate the client.


* fix(dashboard): convert datatable paging through the generated union type

Paging is now a oneOf of numeric enums, so read and write it as JSON
instead of casting the removed scalar type.


---------

+++ internal/kibana/dashboard/panel/aiopschangepointchart/api_test.go
@@ -186,7 +186,7 @@ func TestPopulateFromAPI_nullPreservation(t *testing.T) {
 		AggregationFunction: &agg,
 		SplitField:          new("host.name"),
 		Partitions:          &[]string{"host-a", "host-b"},
-		MaxSeriesToPlot:     new(float32(6)),
+		MaxSeriesToPlot:     new(6),
 		ViewType:            &vt,
 	}
